Refactor MyStepdefs, Add RunnerTest, Add Calculator, Update pom.xml
