+++ src/main/java/calculator/Calculator.java
@@ -1,0 +1,7 @@
+package calculator;
+
+public class Calculator {
+    public int add(int a, int b) {
+        return a + b;
+    }
+}
